% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getInfReps}
\alias{getInfReps}
\title{The code for this function has been taken from the package fishpond, where getInfReps is not an exported function}
\usage{
getInfReps(ys)
}
\description{
The code for this function has been taken from the package fishpond, where getInfReps is not an exported function
}
