% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{infRepError}
\alias{infRepError}
\title{The code for this function has been taken from the package fishpond, where infRepError is not an exported function}
\usage{
infRepError(infRepIdx)
}
\description{
The code for this function has been taken from the package fishpond, where infRepError is not an exported function
}
