% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{trenDi}
\alias{trenDi}
\title{Function is deprecated in favor of mehenDi}
\usage{
trenDi(
  tse,
  x,
  pvalues,
  minP = 0.7,
  mIRVThresh = 0.4,
  alpha = 0.01,
  pThresh = NULL,
  cores = 1
)
}
\arguments{
\item{tse}{TreeSummarizedExperiment which contains the scaled inferential replicates}

\item{x}{character, the name of the condition variable. A factor for two group analysis}

\item{pvalues}{numeric, pvalues for all the nodes in the tree}

\item{minP}{numeric, value betweem 0 and 1, the proportion of the total inferential replicates
which should have the same sign change between conditions (default 0.7)}

\item{mIRVThresh}{numeric, minimum meanInfRV that a node should have for it to be considered for aggregation (default 0.4)}

\item{alpha}{numeric, the rate for the BH correction on the leaves which will be the pvalue threshold for deeming a node significant, wont be used if pThresh is set to some value}

\item{pThresh}{numeric, pvalue threshold for climbing, default is NULL since pThresh is computed automatically based on alpha, if not set to NULL then this pvalue threshold will used for deeming a node significant}

\item{cores}{numeric, the number of cores that will be used during parallelization}
}
\description{
Function is deprecated in favor of mehenDi
}
